feat: support circular targets
The `--circular` option is added.  This will allow a zero cost jump from
the end of the target to the start of the target.

Also clarified the alignment modes further.  Instead of semiglobal,
which is ambiguous, there are now two new alignment modes in its stead.
First is querylocal,  aligns a sub-sequence of the read versus the full
reference.which aligns a sub-sequence of the read,  Next is targetloca,
which aligns the full read versus a sub-sequence of the reference.