feat: speed up traceback significantly and simply circular re-alignment
To speed up traceback, implement a vector-backed hash map for `usize`
keys.  This allows the rapid retrieval, update, and contains operations
for `usize` keys.

Circular re-alignment is simplified, whereby two realignments are
performed for each side of the alignment that ends at the target
boundary.  A first alignment for the aligned portion of the read, and a
second alignment for only the aligned portion that aligns to the target
contig at the boundary.  This obviates the need to re-align each contig
separately to identify origins and perform an expensive multi-contig
re-alignment.